Move create_admin_user, serve_networks and dev_server to hil-admin commands.
* These commands are no longer exposed by the main CLI.
* Changed `serve` to `dev_server` and port is now an optional argument for it.
Port defaults to 5000.
